+++ src/Constants.php
@@ -35,6 +35,10 @@ class Constants extends CommonConstants
      */
     public const DEFAULT_LAPI_URL = 'http://localhost:8080';
     /**
+     * @var string The usage metrics endpoint
+     */
+    public const METRICS_ENDPOINT = '/v1/usage-metrics';
+    /**
      * @var string The user agent prefix used to send request to LAPI
      */
     public const USER_AGENT_PREFIX = 'csphplapi';
@@ -42,4 +46,8 @@ class Constants extends CommonConstants
      * @var string The current version of this library
      */
     public const VERSION = 'v3.3.2';
+    /**
+     * @var string The metrics type
+     */
+    public const METRICS_TYPE = 'crowdsec-php-bouncer';
 }