+++ CHANGELOG.md
@@ -13,6 +13,21 @@ As far as possible, we try to adhere to [Symfony guidelines](https://symfony.com
 
 ---
 
+## [4.0.0](https://github.com/crowdsecurity/php-lapi-client/releases/tag/v4.0.0) - 202?-??-??
+[_Compare with previous release_](https://github.com/crowdsecurity/php-lapi-client/compare/v3.3.2...HEAD)
+
+**This release is not yet published**
+
+### Added
+
+- Add `pushUsageMetrics` method to `Bouncer` class
+
+### Changed
+
+- *Breaking change*: Move configuration classes to `CrowdSec\LapiClient\Configuration` namespace
+
+---
+
 ## [3.3.2](https://github.com/crowdsecurity/php-lapi-client/releases/tag/v3.3.2) - 2024-10-21
 [_Compare with previous release_](https://github.com/crowdsecurity/php-lapi-client/compare/v3.3.1...v3.3.2)